+++ packages/smart-forms-renderer/src/components/FormComponents/ChoiceItems/ChoiceRadioSingle.tsx
@@ -26,10 +26,11 @@ interface ChoiceRadioSingleProps {
   readOnly: boolean;
   disabledViaToggleExpression: boolean;
   fullWidth: boolean;
+  ariaDescribedBy?: string;
 }
 
 function ChoiceRadioSingle(props: ChoiceRadioSingleProps) {
-  const { value, label, readOnly, disabledViaToggleExpression, fullWidth } = props;
+  const { value, label, readOnly, disabledViaToggleExpression, fullWidth, ariaDescribedBy } = props;
 
   const readOnlyVisualStyle = useRendererConfigStore.use.readOnlyVisualStyle();
 
@@ -53,7 +54,14 @@ function ChoiceRadioSingle(props: ChoiceRadioSingleProps) {
       disabled={isHtmlDisabled}
       value={value}
       control={
-        <StandardRadio data-test={`radio-single-${label}`} size="small" readOnly={isHtmlReadOnly} />
+        <StandardRadio
+          data-test={`radio-single-${label}`}
+          size="small"
+          readOnly={isHtmlReadOnly}
+          inputProps={{
+            ...(ariaDescribedBy && { 'aria-describedby': ariaDescribedBy })
+          }}
+        />
       }
       label={label}
     />
