+++ packages/smart-forms-renderer/src/components/FormComponents/TextItem/TextField.tsx
@@ -31,6 +31,7 @@ interface TextFieldProps {
   renderingExtensions: RenderingExtensions;
   readOnly: boolean;
   calcExpUpdated: boolean;
+  instructionsId: string | undefined;
   onInputChange: (value: string) => void;
   onRepopulateSync: (newQrItem: QuestionnaireResponseItem | null) => unknown;
 }
@@ -43,6 +44,7 @@ function TextField(props: TextFieldProps) {
     renderingExtensions,
     readOnly,
     calcExpUpdated,
+    instructionsId,
     onInputChange,
     onRepopulateSync
   } = props;
@@ -66,6 +68,7 @@ function TextField(props: TextFieldProps) {
       slotProps={{
         input: {
           readOnly: readOnly && readOnlyVisualStyle === 'readonly',
+          ...(instructionsId && { 'aria-describedby': instructionsId }),
           endAdornment: (
             <InputAdornment position="end">
               <ExpressionUpdateFadingIcon fadeIn={calcExpUpdated} disabled={readOnly} />
